justeret på diverse projekt-filer så vi får fejlfri build i både x64 og x86 Windiff includerer targetver.h fra stdafx.h

+++ UnixTools/strings/ByteStringExstraftor.cpp
@@ -73,7 +73,7 @@ void ByteStringExtractor::appendChar(char ch) {
 }
 
 void ByteStringExtractor::outputString() {
-  if(m_strLength >= m_minLength) {
+  if(m_strLength >= (int)m_minLength) {
     appendChar('\0');
     printString(m_buf);
   }
